refactor: Outputstream 쓰기 책임 이동
