+++ src/main/java/http/Cookies.java
@@ -1,7 +1,11 @@
 package http;
 
+import java.io.DataOutputStream;
+import java.io.IOException;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import com.google.common.collect.ImmutableSet;
 
@@ -28,5 +32,16 @@ public class Cookies {
     public void putCookie(String key, String value) {
         var cookie = new Cookie(key, value);
         values.add(cookie);
+    }
+
+    public void write(DataOutputStream dos) throws IOException {
+        for (Cookie cookie : values) {
+            var prefix = String.format("Set-Cookie: %s=%s", cookie.getKey(), cookie.getValue());
+
+            var cookieResponse = Stream.concat(Stream.of(prefix), cookie.getOptions().stream())
+                .collect(Collectors.joining("; "));
+
+            dos.writeBytes(cookieResponse + "\r\n");
+        }
     }
 }
